Adds an installer for Logos 10
This changes the install dir to LogosBible10 so it can install alongside verbum or Logos 9. It retains features like appImage for 7.18-staging. And configures the wine prefix for Logos 10. As Logos 10 requires a license the user has to download the installer manually. The installer is the exe now which will always pull the latest msi.