Update controlPanel-Template.sh
Restore --selectAppImage
Change references to --setAppImage to --selectAppImage
wineboot is no longer invoked because it confusingly either fails with no display or if that is removed goes to control panel. It seems nicer to let wineboot happen next time ./Logos.sh is run. Also AppImages are now copied to the /data/bin/ dir rather than the /data/ dir to match the default installation.